Restructure ProjectComponents
