+++ src/Stryker.Core/Stryker.Core/ProjectComponents/Csharp/CsharpFileLeaf.cs
