+++ src/Stryker.Core/Stryker.Core/ProjectComponents/Csharp/CsharpFolderComposite.cs
@@ -1,0 +1,13 @@
+using Microsoft.CodeAnalysis;
+using Stryker.Core.Mutants;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace Stryker.Core.ProjectComponents
+{
+    public class CsharpFolderComposite : FolderComposite<SyntaxTree>
+    {
+
+    }
+}
