--- src/Stryker.Core/Stryker.Core/ProjectComponents/IProjectComponent.cs
@@ -1,22 +1,0 @@
-﻿using Stryker.Core.Mutants;
-using System.Collections.Generic;
-
-namespace Stryker.Core.ProjectComponents
-{
-    public interface IProjectComponent
-    {
-        string FullPath { get; set; }
-        /// <summary>
-        /// Relative path to project file
-        /// </summary>
-        string RelativePath { get; set; }
-
-        IFolderComposite Parent { get; set; }
-
-        IEnumerable<Mutant> Mutants { get; set; }
-
-        IReadOnlyProjectComponent ToReadOnlyInputComponent();
-
-        public abstract IEnumerable<IProjectComponent> GetAllFiles();
-    }
-}
