+++ src/Stryker.Core/Stryker.Core/ProjectComponents/ReadOnly/IReadOnlyProjectComponent.cs
