+++ src/Stryker.Core/Stryker.Core/ProjectComponents/ReadOnly/ReadOnlyFileLeaf.cs
