+++ src/Stryker.Core/Stryker.Core/ProjectComponents/ReadOnly/ReadOnlyFolderComposite.cs
