+++ src/Stryker.Core/Stryker.Core/ProjectComponents/ReadOnly/ReadOnlyProjectComponent.cs
